got lots done on this refactor

+++ KnowYourTurf.Core/Config/CustomPrincipal.cs
@@ -29,8 +29,8 @@ namespace KnowYourTurf.Core.Config
 
         public IIdentity Identity{ get { return _identity; } }
 
-        public long UserId { get{return _userId.IsNotEmpty()? Int64.Parse(_userId):0;} } 
-        public long CompanyId { get { return _companyId.IsNotEmpty()? Int64.Parse(_companyId):0; } }
+        public int UserId { get{return _userId.IsNotEmpty()? Int32.Parse(_userId):0;} } 
+        public int CompanyId { get { return _companyId.IsNotEmpty()? Int32.Parse(_companyId):0; } }
         public string UserRoles { get { return _userRoles; } }
     }
 }
